% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformer.R
\name{data_transformer}
\alias{data_transformer}
\title{Apply SCtransform to a list of Seurat objects}
\usage{
data_transformer(
  object_list,
  verbose = T,
  min_nFeature = 750,
  max_nFeature = 5000,
  max_percent.mt = 7.5
)
}
\arguments{
\item{object_list}{A list of Seurat objects (easily prepared by data_importer function in this package)}

\item{verbose}{Whether to print messages and progress bar for `Seurat::SCTransform()`}

\item{min_nFeature}{Minimum feature number for filtering out cells}

\item{max_nFeature}{Maximum feature number for filtering out cells}

\item{max_percent.mt}{Maximum mitochondrial gene percentage for filtering out cells}
}
\value{
SCtransformed object list
}
\description{
This function is a wrapper to apply SCtransform to a list of Seurat objects.  See Seurat vignette for more details.
}
