% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ranker.R
\name{gene_ranker}
\alias{gene_ranker}
\title{gene_ranker}
\usage{
gene_ranker(
  exprs = NULL,
  pos_marker = NULL,
  neg_marker = NULL,
  sample_id,
  sample_cluster = NULL,
  reference_id,
  reference_cluster = NULL,
  method = "s2n",
  save_ranked_object = F,
  verbose = T
)
}
\arguments{
\item{exprs}{Expression data frame (Single cells are in rows, genes and metadata are in columns)}

\item{pos_marker}{A character vector of gene names to positively gate cells (cells expressing these genes will be included in sample and reference)}

\item{neg_marker}{A character vector of gene names to negatively gate cells (cells expressing these genes will be excluded from sample and reference )}

\item{sample_id}{String. Name of the samples to use for the ranking. As the genes will be ranked from sample to reference in decreasing order, samples will be on the left side of enrichment plots. Regex based string matching can be applied including logic operations such as `|`. Data for this parameter should match to entries under `sample` column from the `expr` data frame.}

\item{sample_cluster}{Which clusters to include in the analysis. Data for this parameter should match to entried under `cluster` column from `expr` data frame.}

\item{reference_id}{Same as `sample_id`, but for reference (right side of the enrichment plots i.e. tail end of the ranking)}

\item{reference_cluster}{Same as sample_cluster but for reference (right side of the enrichment plots)}

\item{method}{Specify method to use for gene ranking. It can be one of the following: 's2n' (default), 'ttest', 'difference', 'ratio', 'welch', 'mwt', 'bws'. See PMID: 18344518}

\item{save_ranked_object}{Logical. Set to TRUE for keeping ranking results as a global variable}

\item{verbose}{Tells you what is being analyzed.}
}
\description{
A function to rank genes between two subsets of scRNAseq data. It uses multiple user-selected statistical approaches. This function works with a simple expression data frame created using `SCseqtools::df_extractor()`` function
}
\examples{
s2n_ranked <- gene_ranker(exprs = exprs,
                          pos_marker = c('Cd3e', 'Cd8a'), neg_marker = 'Cd4',
                          sample_id = 'young|middle_age', reference_id = 'old_age',
                          method = 's2n')



}
