% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGseaTable2.R
\name{plotGseaTable2}
\alias{plotGseaTable2}
\title{plotGseaTable2}
\usage{
plotGseaTable2(
  pathways,
  stats,
  fgseaRes,
  gseaParam = 1,
  colwidths = c(5, 3, 0.8, 1.2, 1.2),
  plot_title = NULL,
  do.plot = T
)
}
\arguments{
\item{pathways}{Pathways to plot table, as in 'fgsea' function. List object containing reference gene sets for analyses. Reference pathways correspond to name of the list item.}

\item{stats}{Gene-level stats, as in 'fgsea' function. A named numeric vector of ranked genes.}

\item{fgseaRes}{Table with fgsea results. A previously calculated data frame containing gsea results}

\item{gseaParam}{GSEA-like parameter. Adjusts displayed statistic values, values closer to 0 flatten plots. Default = 1, value of 0.5 is a good choice too. Numeric value to increase the bar size of the genes in enrichment plot.}

\item{colwidths}{Vector of five elements corresponding to column width for grid.arrange. If column width is set to zero, the column is not drawn.}

\item{plot_title}{Add custom title to the summary plots. It can be useful for generating plots in a loop}

\item{do.plot}{Logical. Set it to TRUE to create graphical outputs.}
}
\description{
This function plots a GSEA results. The function was edited from fgsea package to enable to graph naming
}
