% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_plotter.R
\name{top_plotter}
\alias{top_plotter}
\title{top_plotter}
\usage{
top_plotter(
  gsea_results = NULL,
  ranked_genes = NULL,
  gene_set = NULL,
  top_n = 10,
  gseaParam = 1,
  plot_title = NULL,
  do.plot = T
)
}
\arguments{
\item{gsea_results}{A previously calculated data frame containing gsea results}

\item{ranked_genes}{A named numeric vector of ranked genes.}

\item{gene_set}{List object containing reference gene sets for analyses. Reference pathways correspond to name of the list item.}

\item{top_n}{Specify the number of top pathways to be shown}

\item{gseaParam}{Numeric value to increase the bar size of the genes in enrichment plot}

\item{plot_title}{Add custom title to the summary plots. It can be useful for generating plots in a loop}

\item{do.plot}{Logical. Set it to TRUE to create graphical outputs.}
}
\description{
A function to speed up plotting GSEA results. Use this function with previously generated "gene rankings" and "gsea results"
}
\examples{

top_plotter(gsea_results = f480_d12_gsea, ranked_genes = f480_d12_rnk, gene_set = msig_HALLMARK, top_n = #' 5, gseaParam = 0.4)

top_plotter(gsea_results = act_cd8_d12_gsea, ranked_genes = act_cd8_d12_rnk, gene_set = msig_HALLMARK, #' top_n = 5, gseaParam = 0.4)



}
