% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_plotter.R
\name{gsea_plotter}
\alias{gsea_plotter}
\title{gsea_plotter}
\usage{
gsea_plotter(
  exprs = NULL,
  preranked_genes = NULL,
  pos_marker = NULL,
  neg_marker = NULL,
  sample_id = NULL,
  sample_cluster = NULL,
  reference_id = NULL,
  reference_cluster = NULL,
  method = "s2n",
  gene_set = "hallmark",
  nperm = 10000,
  minSize = 50,
  maxSize = 500,
  top_n = 10,
  gseaParam = 1,
  plot_individual = NULL,
  append_title = F,
  top_plots_title = T,
  seed = 123,
  keep_results = T,
  save_png = F,
  png_units = "in",
  png_width = 4,
  png_height = 3,
  append_to_filename = "",
  verbose = T,
  annot_text_color = "black",
  annot_text_size = 4,
  annot_text_fontface = 2
)
}
\arguments{
\item{exprs}{}

\item{preranked_genes}{}

\item{pos_marker}{A character vector of gene names to positively gate cells (cells expressing these genes will be included in sample and reference)}

\item{neg_marker}{A character vector of gene names to negatively gate cells (cells expressing these genes will be excluded from sample and reference )}

\item{sample_id}{String. Name of the samples to use for the ranking. As the genes will be ranked from sample to reference in decreasing order, samples will be on the left side of enrichment plots. Regex based string matching can be applied including logic operations such as `|`. Data for this parameter should match to entries under `sample` column from the `expr` data frame.}

\item{sample_cluster}{Which clusters to include in the analysis. Data for this parameter should match to entried under `cluster` column from `expr` data frame.}

\item{reference_id}{Same as `sample_id`, but for reference (right side of the enrichment plots i.e. tail end of the ranking)}

\item{reference_cluster}{Same as sample_cluster but for reference (right side of the enrichment plots)}

\item{method}{Specify method to use for gene ranking. It can be one of the following: 's2n' (default), 'ttest', 'difference', 'ratio', 'welch', 'mwt', 'bws'. See PMID: 18344518}

\item{gene_set}{reference pathway sets. a list object}

\item{nperm}{The number of permutations to use in analysis}

\item{minSize}{Minimum number of overlapping genes to consider a pathway in analysis}

\item{maxSize}{Maximum number of overallping genes to consider a pathway in analysis}

\item{top_n}{Specificy how many of the top enriched pathways are shown}

\item{gseaParam}{fgsea parameter to change bar sizes}

\item{plot_individual}{Select pathway to plot}

\item{append_title}{Add informative titles to individual plots}

\item{seed}{Random seed}

\item{keep_results}{Logical to store results globally}

\item{save_png}{Save resulting plot as png}

\item{png_units}{Png size units}

\item{png_width}{width}

\item{png_height}{height}

\item{append_to_filename}{Custom string to add to png file}

\item{verbose}{Prints the analysis details}

\item{annot_text_color}{color of NES-p val annotation text}

\item{annot_text_size}{size of NES-pval annotation text}

\item{annot_text_fontface}{fontface of annotation text (1,2,3,4, plain-bold-italic-bold and italic)}

\item{top_plots_totle}{Add informative titles to summary plots}
}
\description{
A wrapper function to create gsea plots.
}
